some bugs have been debuged. after version4.1  Changes to be committed: 	new file:   1_Visual/1_Visual-1-1.edat2 	new file:   1_Visual/1_Visual-1-1.txt 	new file:   1_Visual/1_Visual.ebs2 	new file:   1_Visual/1_Visual.es2 	new file:   1_Visual/1_Visual.wndpos 	new file:   1_Visual/Dummy.JPG 	new file:   1_Visual/PreScan.JPG 	new file:   1_Visual/Visual/Slide1.JPG 	new file:   1_Visual/Visual/Slide2.JPG 	new file:   1_Visual/Visual/Slide3.JPG 	new file:   1_Visual/Visual/Slide3.bmp 	new file:   1_Visual/Visual/Slide4.JPG 	new file:   1_Visual/Visual/Slide4.bmp 	new file:   1_Visual/Visual/Thumbs.db 	new file:   1_Visual/Visual/Visual.pptx 	new file:   1_Visual/myBlockList.txt 	new file:   1_Visual/myTrialListParameters.txt 	new file:   1_Visual/otherParameters.txt 	modified:   smartfMRI/GeneratedFiles/qrc_smartfmri.cpp 	modified:   smartfMRI/license.dat 	modified:   smartfMRI/logfiledata.cpp 	new file:   smartfMRI/paradigm/!@#$%^/1_Visual-1-1.edat2 	new file:   smartfMRI/paradigm/!@#$%^/1_Visual-1-1.txt 	new file:   smartfMRI/paradigm/!@#$%^/1_Visual.ebs2 	new file:   smartfMRI/paradigm/!@#$%^/1_Visual.es2 	new file:   smartfMRI/paradigm/!@#$%^/1_Visual.wndpos 	new file:   smartfMRI/paradigm/!@#$%^/Dummy.JPG 	new file:   smartfMRI/paradigm/!@#$%^/PreScan.JPG 	new file:   smartfMRI/paradigm/!@#$%^/Visual/Slide1.JPG 	new file:   smartfMRI/paradigm/!@#$%^/Visual/Slide2.JPG 	new file:   smartfMRI/paradigm/!@#$%^/Visual/Slide3.JPG 	new file:   smartfMRI/paradigm/!@#$%^/Visual/Slide3.bmp 	new file:   smartfMRI/paradigm/!@#$%^/Visual/Slide4.JPG 	new file:   smartfMRI/paradigm/!@#$%^/Visual/Slide4.bmp 	new file:   smartfMRI/paradigm/!@#$%^/Visual/Visual.pptx 	new file:   smartfMRI/paradigm/!@#$%^/myBlockList.txt 	new file:   smartfMRI/paradigm/!@#$%^/myTrialListParameters.txt 	new file:   smartfMRI/paradigm/!@#$%^/otherParameters.txt 	new file:   smartfMRI/paradigm/PictureRT_Test2/BlueCar.bmp 	new file:   smartfMRI/paradigm/PictureRT_Test2/PictureRT-1-1-ExperimentAdvisorReport.xml 	new file:   smartfMRI/paradigm/PictureRT_Test2/PictureRT-1-1.edat2 	new file:   smartfMRI/paradigm/PictureRT_Test2/PictureRT-1-1.txt 	new file:   smartfMRI/paradigm/PictureRT_Test2/PictureRT.ebs2 	new file:   smartfMRI/paradigm/PictureRT_Test2/PictureRT.es2 	new file:   smartfMRI/paradigm/PictureRT_Test2/PictureRT.wndpos 	new file:   smartfMRI/paradigm/PictureRT_Test2/RedCar.bmp 	new file:   smartfMRI/paradigm/PictureRT_Test2/advisor.xsl 	new file:   smartfMRI/paradigm/PictureRT_Test2/log/PictureRT20160807212031/PictureRT20160807212031.edat2 	new file:   smartfMRI/paradigm/PictureRT_Test2/log/PictureRT20160807212031/PictureRT20160807212031.txt 	new file:   smartfMRI/paradigm/PictureRT_Test2/log/PictureRT20160807213424/PictureRT20160807213424.edat2 	new file:   smartfMRI/paradigm/PictureRT_Test2/log/PictureRT20160807213424/PictureRT20160807213424.txt 	new file:   smartfMRI/paradigm/PictureRT_Test2/log/PictureRT20160807213735/PictureRT20160807213735.edat2 	new file:   smartfMRI/paradigm/PictureRT_Test2/log/PictureRT20160807213735/PictureRT20160807213735.txt 	new file:   smartfMRI/paradigm/PictureRT_Test2/log/PictureRT20160807224244/PictureRT20160807224244.edat2 	new file:   smartfMRI/paradigm/PictureRT_Test2/log/PictureRT20160807224244/PictureRT20160807224244.txt 	new file:   smartfMRI/paradigm/PictureRT_Test2/log/PictureRT20160807225958/PictureRT20160807225958.edat2 	new file:   smartfMRI/paradigm/PictureRT_Test2/log/PictureRT20160807225958/PictureRT20160807225958.txt 	new file:   smartfMRI/paradigm/PictureRT_Test2/myBlockList.txt 	new file:   smartfMRI/paradigm/PictureRT_Test2/myTrialListParameters.txt 	new file:   smartfMRI/paradigm/PictureRT_Test2/otherParameters.txt 	modified:   smartfMRI/qcopydirrecursively.cpp 	modified:   smartfMRI/qcopydirrecursively.h 	modified:   smartfMRI/smartfmri.cpp 	modified:   smartfMRI/statusthread.cpp

+++ smartfMRI/smartfmri.cpp
@@ -1,5 +1,7 @@
 #include "smartfmri.h"
 
+#include <QMessageBox>
+
 SmartfMRI::SmartfMRI(QWidget *parent)
 	: QMainWindow(parent), expMan("./paradigm", parent)
 {
@@ -30,27 +32,37 @@ int SmartfMRI::removeExperiment()
 	QDir dirToBeDeleted(ui.experimentlistView->
 		currentIndex().data(Qt::ToolTipRole).toString());
 	dirToBeDeleted.setFilter(QDir::NoDotAndDotDot);
-	if (ui.experimentlistView->currentIndex().data().isValid() 
-		&& dirToBeDeleted.removeRecursively()) {
-		if (expMod != nullptr) {
-			delete expMod;
-		}
-		expMod = new ExperimentModel(QDir("./paradigm"), this);
-		ui.experimentlistView->setModel(expMod);
+	QMessageBox sureToDelete;
+	sureToDelete.setText("Are you sure to remove this Experiment?\n"
+		"It will remove all data(including paradigm, parameters, log, etc)"
+		" in this experiment folder and cannot be recovered!");
+	sureToDelete.setStandardButtons(QMessageBox::Yes | QMessageBox::No);
+	sureToDelete.setDefaultButton(QMessageBox::No);
+	
+	if (sureToDelete.exec() == QMessageBox::Yes) {
+		if (ui.experimentlistView->currentIndex().data().isValid()
+			&& dirToBeDeleted.removeRecursively()) {
+			if (expMod != nullptr) {
+				delete expMod;
+			}
+			expMod = new ExperimentModel(QDir("./paradigm"), this);
+			ui.experimentlistView->setModel(expMod);
 
-		qDebug() << "remove successfully";
-		return 1;
+			qDebug() << "remove successfully";
+			return 1;
+		}
+		else if (!ui.experimentlistView->currentIndex().data().isValid()) {
+			QMessageBox::critical(this, tr("Fail to remove"),
+				QString("Please select an item."));
+			return 0;
+		}
+		else {
+			QMessageBox::critical(this, tr("Fail to remove"),
+				QString("Please remove ") + dirToBeDeleted.absolutePath() + " by yourself");
+			return 0;
+		}
 	}
-	else if (!ui.experimentlistView->currentIndex().data().isValid()) {
-		QMessageBox::critical(this, tr("Fail to remove"),
-			QString("Please select an item."));
-		return 0;
-	}
-	else {
-		QMessageBox::critical(this, tr("Fail to remove"),
-			QString("Please remove ") + dirToBeDeleted.absolutePath() + " by yourself");
-		return 0;
-	}
+	return 0;
 }
 
 int SmartfMRI::addExperiment() {
@@ -143,7 +155,12 @@ int SmartfMRI::openLogFolder()
 	}
 	else {
 		Experiment* e = expMod->getExperiment(ui.experimentlistView->currentIndex().data().toString());
-		QDesktopServices::openUrl(QUrl::fromLocalFile(e->getDir().absolutePath() + "/log"));
+		qDebug() << "Open log folder";
+		if (!QDesktopServices::openUrl(
+			QUrl::fromLocalFile(e->getDir().absolutePath() + "/log"))) {
+			QMessageBox::critical(this, tr("Fail to open"),
+				QString("No Log files has been created yet"));
+		}
 		return 1;
 	}
